+++ cli/docs/changelog.md
@@ -1,5 +1,11 @@
 # TBD
 
+# 0.18.3
+### Fixes
+* Upgrade to latest container-engine-lib, Core, and Engine which contain fixes for modules on Kubernetes
+* Switched `module exec` to the new Kurtosis command framework
+* Fixed the logs `ReadCloser`s returned by `service logs` and `module exec` not getting closed
+
 # 0.18.2
 ### Fixes
 * Fix broken `service logs` command